requirements.txt
nats-py
opencv-python
numpy
